+++ CHANGELOG.md
@@ -49,6 +49,8 @@
 - `general`
   - [#1525](https://github.com/wix-incubator/rich-content/pull/1525) typed theme functions for all plugins
   - [#1528](https://github.com/wix-incubator/rich-content/pull/1528) fixed `npm run watch` command for several packages
+- `storybook`
+  - [#1533](https://github.com/wix-incubator/rich-content/pull/1533) ricos theme's palette examples are more realistic
 
 </details>